+++ email-worker/src/common/dependency-symbols.ts
@@ -1,3 +1,3 @@
 export const DEPENDENCY_SYMBOLS = {
   Notifier: Symbol.for('Notifier'),
-} as const;
+};
