+++ email-worker/src/common/types.ts
@@ -23,4 +23,3 @@ export interface RssRemoval {
   rssUrl: string;
   certificateCode: string;
 }
-
