+++ email-worker/src/email/types.ts
@@ -1,4 +1,4 @@
-import { RssRegistration, RssRemoval, User } from '@app-types/types';
+import { RssRegistration, RssRemoval, User } from '@common/types';
 
 import { EmailPayloadConstant } from './constant';
 
@@ -11,3 +11,10 @@ export type EmailPayload =
     }
   | { type: typeof EmailPayloadConstant.PASSWORD_RESET; data: User }
   | { type: typeof EmailPayloadConstant.ACCOUNT_DELETION; data: User };
+
+export type NodeMailerError = Error & {
+  code?: string;
+  command?: string;
+  response?: string;
+  responseCode?: number;
+};
