--- email-worker/src/types/nodemailer.ts
@@ -1,6 +1,0 @@
-export interface NodeMailerError extends Error {
-  code?: string;
-  command?: string;
-  response?: string;
-  responseCode?: number;
-}
