+++ email-worker/test/config/e2e/common/rabbitmq-test.helper.ts
@@ -1,7 +1,7 @@
 import { Channel } from 'amqplib';
 import { StartedTestContainer } from 'testcontainers';
 
-import { EmailPayload } from '@email/type';
+import { EmailPayload } from '@email/types';
 
 import {
   RMQ_EXCHANGES,
