+++ email-worker/test/config/e2e/common/testContext.setup.ts
@@ -1,15 +1,15 @@
 import { container, DependencyContainer } from 'tsyringe';
 
-import { DiscordNotifier } from '@notification/discord.notifier';
-import { Notifier } from '@notification/notifier.interface';
+import { DEPENDENCY_SYMBOLS } from '@common/dependency-symbols';
 
 import { EmailConsumer } from '@email/email.consumer';
 import { EmailService } from '@email/email.service';
 
+import { DiscordNotifier } from '@notification/discord.notifier';
+import { Notifier } from '@notification/notifier.interface';
+
 import { RabbitMQManager } from '@rabbitmq/rabbitmq.manager';
 import { RabbitMQService } from '@rabbitmq/rabbitmq.service';
-
-import { DEPENDENCY_SYMBOLS } from '@app-types/dependency-symbols';
 
 export interface TestContext {
   container: DependencyContainer;
